+++ pages/signup.tsx
@@ -1,6 +1,6 @@
 import Image from "next/image";
-import { useRouter } from "next/router";
-import { useEffect, useState } from "react";
+import router from "next/router";
+import { useEffect } from "react";
 import Logo from "src/assets/icon/logo.svg";
 import Button from "src/components/Button/Button";
 import Input from "src/components/Input/Input";
@@ -8,48 +8,18 @@ import InputLabel from "src/components/InputLabel/InputLabel";
 import SignText from "src/components/SignText/SignText";
 import SocialLogin from "src/components/SocialLogin/SocialLogin";
 import { INPUT_TYPE } from "src/constants/input";
-import { checkToken, getToken } from "src/utils/aboutToken";
+import { useSignup } from "src/hook/SignUp/useSignup";
+import { checkToken } from "src/utils/aboutToken";
 import styles from "styles/SignPage.module.css";
-import { postDuplicationEmail, postSignUp } from "./api/api";
 
 function SignUp() {
   const { email, password, passwordCheck } = INPUT_TYPE;
-  const [status, setStatus] = useState(0);
-  const router = useRouter();
-  const [account, setAccount] = useState({
-    email: "",
-    password: "",
-    passwordCheck: "",
-  });
-
-  const handleSignUp = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    setStatus(0);
-    const signUp = await postSignUp(account);
-
-    if (signUp.error) {
-      setStatus(400);
-    }
-
-    if (signUp.data && account.password === account.passwordCheck) {
-      getToken("access-token", signUp.data.accessToken);
-      router.push("/folder");
-    }
-  };
+  const { account, setAccount, status, setStatus, handleSignUp } = useSignup();
 
   const handleEnter = (e: React.KeyboardEvent<HTMLFormElement>) => {
     if (e.key === "Enter") {
       handleSignUp(e);
       setStatus(0);
-    }
-  };
-
-  const handleDuplicateCheck = async (e: React.MouseEvent) => {
-    e.preventDefault();
-    setStatus(0);
-    const duplicateEmail = await postDuplicationEmail(account);
-    if (duplicateEmail.error.message === email.duplicateEmail) {
-      setStatus(409);
     }
   };
 
@@ -81,12 +51,6 @@ function SignUp() {
               account={account}
               setAccount={setAccount}
             />
-            <button
-              className={styles.emailCheck}
-              onClick={handleDuplicateCheck}
-            >
-              중복 체크
-            </button>
           </div>
           <div className={styles.passwordInput}>
             <InputLabel htmlFor={password.type} content="비밀번호" />
